draft full workflow
