+++ package.json
@@ -8,7 +8,6 @@
     "build": "run-s build:*",
     "build:drizzle": "npm run db:generate",
     "build:react-router": "react-router build",
-    "build:docs": "tsx ./scripts/generate-openapi.ts",
     "start": "cross-env NODE_ENV=production react-router-serve ./build/server/index.js",
     "test": "vitest",
     "test:ui": "vitest --ui",
@@ -121,7 +120,6 @@
     "remix-i18next": "^7.5.0",
     "rollup-preserve-directives": "^1.1.3",
     "simple-statistics": "^7.8.9",
-    "swagger-jsdoc": "^6.2.8",
     "swagger-ui-react": "^5.32.4",
     "tailwind-merge": "^3.5.0",
     "tailwindcss-animate": "^1.0.7",
@@ -156,7 +154,6 @@
     "@types/react": "^19.2.14",
     "@types/react-dom": "19.2.3",
     "@types/supercluster": "^7.1.3",
-    "@types/swagger-jsdoc": "^6.0.4",
     "@types/swagger-ui-react": "^5.18.0",
     "@vitejs/plugin-react": "^6.0.1",
     "@vitest/coverage-v8": "^4.1.5",
